requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
psutil==5.9.8
python-dotenv==1.0.1
httpx==0.27.0
redis==5.0.1

